[lexers] Improve the sententence building algorithm.
* Now handles a single letter followed by a period as a bad place to end a sentence.
* Introduces abbreviations to handle 'Mr.', 'Mrs.' etc.

It's worth to note that the algorithm is 'greedy' in that if there's any
uncertainty if a sentence should end or not at a certain point, it doesn't.

The goal of the lexer is not to perfectly break the text into individual
sentences but rather guarantee that the text is broken into blocks of text
containing one or several complete sentences.