[lexers/tests] Change the way comments are handled.
A comment is now stripped of any leading whitespace and every
character up until EOF/newline is emitted as one whole token.
The idea is that the parser will attach a special 'comment' scope
to the token and emit the comment string as its own segment. The
linter will ignore comments by default unless a rule specifically
targets them.