[parsers/tests] Don't allow option enclosures in math scopes.
The characters '[' and ']' are much more likely to be used to
specify ranges rather than options. In order to avoid parsing
problems with constructions as "x \in [0.5, 1)" we simply
disable option enclosures in math scopes. Time will tell if
something more complex is needed or if this naive solition
is enough in practice.